// @hey_api/documentsmaterials.swagger/types.gen.ts
export type v2Category = {
  /**
   * Unique identifier of the category.
   */
  categoryId?: string;
  /**
   * Titles of the category in different languages.
   */
  titles?: v2Titles;
  /**
   * Labels assigned to the category.
   */
  labels?: Array<string>;
  /**
   * Multilingual description of the category.
   */
  description?: v2DescriptionJSON;
  /**
   * ID of the user who created the category.
   */
  userId?: string;
  /**
   * number of documents linked to this specific category.
   */
  documentCount?: number;
  /**
   * Timestamp of category creation.
   */
  createdAt?: string;
  /**
   * Timestamp of last update.
   */
  updatedAt?: string;
};
// ...
export type v2DescriptionJSON = {
  en?: string;
  fr?: string;
};
// ...
export type v2Titles = {
  en?: string;
  fr?: string;
};